requirements/local.txt
-r base.txt
django-debug-toolbar==1.6
ipython==5.3.0

